backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
fastapi-mail==1.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
slowapi==0.1.9
bleach==6.2.0
alembic==1.13.1
